+++ canopen_sniffer/dox/documentation/html/search/all_10.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['time_101',['TIME',['../classcanopen__sniffer_1_1frame__type.html#a256d466ef5a544c9fad1adb69ce28ca8',1,'canopen_sniffer::frame_type']]],
-  ['total_102',['total',['../classcanopen__sniffer_1_1bus__stats_1_1frame__count.html#a2c1a52cd37ed38e861bba51c47c1802a',1,'canopen_sniffer::bus_stats::frame_count']]]
+  ['time_102',['TIME',['../classcanopen__sniffer_1_1frame__type.html#a256d466ef5a544c9fad1adb69ce28ca8',1,'canopen_sniffer::frame_type']]],
+  ['total_103',['total',['../classcanopen__sniffer_1_1bus__stats_1_1frame__count.html#a2c1a52cd37ed38e861bba51c47c1802a',1,'canopen_sniffer::bus_stats::frame_count']]]
 ];