+++ canopen_sniffer/dox/documentation/html/search/all_b.js
@@ -1,5 +1,6 @@
 var searchData=
 [
   ['main_65',['main',['../namespacecanopen__sniffer.html#a355f06d5ba7ea264dc3e2e128349c130',1,'canopen_sniffer']]],
-  ['max_5fframes_66',['MAX_FRAMES',['../namespacecanopen__sniffer.html#a90270d61ef5cbcf52b1f4e63ee96cfbc',1,'canopen_sniffer']]]
+  ['max_5fframes_66',['MAX_FRAMES',['../namespacecanopen__sniffer.html#a90270d61ef5cbcf52b1f4e63ee96cfbc',1,'canopen_sniffer']]],
+  ['max_5fstats_5fshow_67',['MAX_STATS_SHOW',['../namespacecanopen__sniffer.html#aa6fda8207dcce3b56fd8ed0ff3ec827e',1,'canopen_sniffer']]]
 ];