+++ canopen_sniffer/dox/documentation/html/search/classes_2.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['displaycli_109',['DisplayCLI',['../classcanopen__sniffer_1_1DisplayCLI.html',1,'canopen_sniffer']]],
-  ['displaygui_110',['DisplayGUI',['../classcanopen__sniffer_1_1DisplayGUI.html',1,'canopen_sniffer']]]
+  ['displaycli_110',['DisplayCLI',['../classcanopen__sniffer_1_1DisplayCLI.html',1,'canopen_sniffer']]],
+  ['displaygui_111',['DisplayGUI',['../classcanopen__sniffer_1_1DisplayGUI.html',1,'canopen_sniffer']]]
 ];