+++ canopen_sniffer/dox/documentation/html/search/functions_1.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['add_5fnode_124',['add_node',['../classcanopen__sniffer_1_1bus__stats.html#ac8f70966bf334462b64f0bcf6e1f7474',1,'canopen_sniffer::bus_stats']]]
+  ['add_5fnode_125',['add_node',['../classcanopen__sniffer_1_1bus__stats.html#ac8f70966bf334462b64f0bcf6e1f7474',1,'canopen_sniffer::bus_stats']]]
 ];