+++ canopen_sniffer/dox/documentation/html/search/variables_6.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['graph_5fwidth_186',['GRAPH_WIDTH',['../classcanopen__sniffer_1_1DisplayCLI.html#a3f49a3f3cb41debc86ca3d71a7ad500c',1,'canopen_sniffer::DisplayCLI']]]
+  ['graph_5fwidth_187',['GRAPH_WIDTH',['../classcanopen__sniffer_1_1DisplayCLI.html#a3f49a3f3cb41debc86ca3d71a7ad500c',1,'canopen_sniffer::DisplayCLI']]]
 ];