+++ canopen_sniffer/dox/documentation/html/search/variables_a.js
@@ -1,4 +1,5 @@
 var searchData=
 [
-  ['max_5fframes_193',['MAX_FRAMES',['../namespacecanopen__sniffer.html#a90270d61ef5cbcf52b1f4e63ee96cfbc',1,'canopen_sniffer']]]
+  ['max_5fframes_194',['MAX_FRAMES',['../namespacecanopen__sniffer.html#a90270d61ef5cbcf52b1f4e63ee96cfbc',1,'canopen_sniffer']]],
+  ['max_5fstats_5fshow_195',['MAX_STATS_SHOW',['../namespacecanopen__sniffer.html#aa6fda8207dcce3b56fd8ed0ff3ec827e',1,'canopen_sniffer']]]
 ];