+++ canopen_sniffer/dox/documentation/html/search/variables_e.js
@@ -1,9 +1,9 @@
 var searchData=
 [
-  ['sdo_5fframes_205',['sdo_frames',['../classcanopen__sniffer_1_1DisplayCLI.html#a5b5c93be2f940dcfa114fa61f98c1d23',1,'canopen_sniffer::DisplayCLI']]],
-  ['sdo_5freq_206',['SDO_REQ',['../classcanopen__sniffer_1_1frame__type.html#aa227b2666e9ee4b630add48d6dec0c39',1,'canopen_sniffer::frame_type']]],
-  ['sdo_5fres_207',['SDO_RES',['../classcanopen__sniffer_1_1frame__type.html#a95373d5028d5cb3c1ce7b6120de91447',1,'canopen_sniffer::frame_type']]],
-  ['set_208',['set',['../classcanopen__sniffer_1_1bus__stats_1_1stats__data.html#a112a711f766db446c2f517b794a1c04a',1,'canopen_sniffer::bus_stats::stats_data']]],
-  ['stats_209',['stats',['../classcanopen__sniffer_1_1process__frame.html#aa57ebe01934de43865125819a3c4af74',1,'canopen_sniffer.process_frame.stats()'],['../classcanopen__sniffer_1_1DisplayCLI.html#aa57ebe01934de43865125819a3c4af74',1,'canopen_sniffer.DisplayCLI.stats()']]],
-  ['sync_210',['SYNC',['../classcanopen__sniffer_1_1frame__type.html#a03082caece11e4751c1e251f4a2525cd',1,'canopen_sniffer::frame_type']]]
+  ['sdo_5fframes_207',['sdo_frames',['../classcanopen__sniffer_1_1DisplayCLI.html#a5b5c93be2f940dcfa114fa61f98c1d23',1,'canopen_sniffer::DisplayCLI']]],
+  ['sdo_5freq_208',['SDO_REQ',['../classcanopen__sniffer_1_1frame__type.html#aa227b2666e9ee4b630add48d6dec0c39',1,'canopen_sniffer::frame_type']]],
+  ['sdo_5fres_209',['SDO_RES',['../classcanopen__sniffer_1_1frame__type.html#a95373d5028d5cb3c1ce7b6120de91447',1,'canopen_sniffer::frame_type']]],
+  ['set_210',['set',['../classcanopen__sniffer_1_1bus__stats_1_1stats__data.html#a112a711f766db446c2f517b794a1c04a',1,'canopen_sniffer::bus_stats::stats_data']]],
+  ['stats_211',['stats',['../classcanopen__sniffer_1_1process__frame.html#aa57ebe01934de43865125819a3c4af74',1,'canopen_sniffer.process_frame.stats()'],['../classcanopen__sniffer_1_1DisplayCLI.html#aa57ebe01934de43865125819a3c4af74',1,'canopen_sniffer.DisplayCLI.stats()']]],
+  ['sync_212',['SYNC',['../classcanopen__sniffer_1_1frame__type.html#a03082caece11e4751c1e251f4a2525cd',1,'canopen_sniffer::frame_type']]]
 ];